vuejs/todoapp/requirements.txt
coverage==4.4.1
coveralls==1.1
Django==1.10
django-nose==1.4.4
djangorestframework==3.6.3
pytz==2017.2

